+++ src/slack/messenger.ts
@@ -8,9 +8,11 @@ import {
   CurrentUser,
   Users,
   MessageReply,
-  Providers
+  Providers,
+  UserPresence
 } from "../types";
 import { SelfCommands } from "../constants";
+import { IDNDStatus } from "./common";
 
 const RTMEvents = {
   AUTHENTICATED: "authenticated",
@@ -21,7 +23,9 @@ const RTMEvents = {
   PRESENCE_CHANGE: "presence_change",
   CHANNEL_MARKED: "channel_marked",
   GROUP_MARKED: "group_marked",
-  IM_MARKED: "im_marked"
+  IM_MARKED: "im_marked",
+  DND_UPDATED_SELF: "dnd_updated",
+  DND_UPDATED_USER: "dnd_updated_user"
 };
 
 const EventSubTypes = {
@@ -33,7 +37,11 @@ const EventSubTypes = {
 class SlackMessenger {
   rtmClient: RTMClient;
 
-  constructor(private token: string) {
+  constructor(
+    token: string,
+    private onPresenceChanged: any,
+    private onDndStateChanged: any
+  ) {
     // We can also use { useRtmConnect: false } for rtm.start
     // instead of rtm.connect, which has more fields in the payload
     let options: RTMClientOptions = {};
@@ -129,15 +137,6 @@ class SlackMessenger {
       });
     });
 
-    this.rtmClient.on(RTMEvents.PRESENCE_CHANGE, event => {
-      const { user: userId, presence } = event;
-      const isOnline = presence === "active";
-      vscode.commands.executeCommand(SelfCommands.UPDATE_USER_PRESENCE, {
-        userId,
-        isOnline
-      });
-    });
-
     this.rtmClient.on(RTMEvents.CHANNEL_MARKED, event => {
       const { channel, ts, unread_count_display } = event;
       vscode.commands.executeCommand(SelfCommands.CHANNEL_MARKED, {
@@ -163,6 +162,25 @@ class SlackMessenger {
         readTimestamp: ts,
         unreadCount: unread_count_display
       });
+    });
+
+    this.rtmClient.on(RTMEvents.PRESENCE_CHANGE, event => {
+      const { user: userId, presence: rawPresence } = event;
+      this.onPresenceChanged(userId, rawPresence);
+    });
+
+    this.rtmClient.on(RTMEvents.DND_UPDATED_SELF, event => {
+      // This is a no-op, since the DND_UPDATED_USER handler does everything
+      const { user: userId } = event;
+      const dndStatus: IDNDStatus = event.dnd_status;
+    });
+
+    this.rtmClient.on(RTMEvents.DND_UPDATED_USER, event => {
+      // This event is not fired when the snooze ends for a user
+      // Hence, we need to maintain that via dnd timers
+      const { user: userId } = event;
+      const dndStatus: IDNDStatus = event.dnd_status;
+      this.onDndStateChanged(userId, dndStatus);
     });
   }